refactor: domain 분리

+++ src/main/java/com/goormthon/bookduchilseong/domain/bookclub/dto/request/BookClubTogetherRequestDTO.java
@@ -19,5 +19,5 @@ public class BookClubTogetherRequestDTO {
 	private String profile;
 	private String bookTitle;
 	private String author;
-	private int totalPage;
+	private Long totalPage;
 }
